+++ src/main/java/com/josiahebhomenye/sound/SquareWave.java
@@ -14,9 +14,9 @@ public class SquareWave extends SoundGenerator implements WaveFunctions {
     @Override
     public void generate() {
         for(int i = 0; i < nSamples; i++){
-            float t = timeStep * i;
-            float sample = 0;
-            for(float n = 1; n <= limit; n++){
+            double t = timeStep * i;
+            double sample = 0;
+            for(double n = 1; n <= limit; n++){
                 sample += (1.0/(2 * n - 1)) * sin((2 * n - 1) * t);
             }
             channels.write(sample);
